+++ install/linux/tlp.sh
@@ -3,6 +3,11 @@
 set -Eeuo pipefail
 
 
+if dpkg -s tlp &> /dev/null; then
+    echo "tlp is already installed; exiting"
+    exit 0
+fi
+
 sudo apt install -y tlp
 sudo systemctl enable tlp.service && sudo tlp start
